Second version of a boid
Work in Progress

BrainyBoid:
This boid is meant to have a neural network to direct its movements. It will have vision of its surroundings. 

Environment is represented by a class. Environment will have a list of objects within it. Display environment will be done using PIL and matplotlib to have fixed colours for different objects.